requirements.txt
django
pymongo
mongoengine
djangorestframework
drf-yasg
requests
django-environ